# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API for users and contents",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/delete-user/{id}": {
            "delete": {
                "description": "Delete selected author's data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Unacceptable"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update-author/{id}": {
            "put": {
                "description": "Update selected author's data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Author",
                        "in": "body",
                        "description": "Author's Data",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Unacceptable"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/edit-author-rest/{id}": {
            "get": {
                "description": "Return edit admin page view",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Unacceptable"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin-rest": {
            "get": {
                "description": "Return admin page view",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Unacceptable"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/excel-export": {
            "get": {
                "description": "Return excel report page view",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Unacceptable"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {}
}